Added creature object pool
Changed creature list to incorporate a creature pool. Now when a creature dies it is moved to the pool and when a new one is created it is taken from the pool and reset. Due to the high volume of creatures dieing and being born this should have prvoided a certain amount of performance gains, however this has not proved to be the case. I'm leaving the pool regardless as it at least will help with memory fragmentation and may provide performance gains when things are scaled up.

Interesting point - when pregnant creatures were dying in previous versions their babies were not being deleted causing a memory leak I wasn't aware of.