Added save load creature from file & carnivores
Added the function to save all the current creatures to a file with a keyboard command.
Added a function to load all the cretures in a file and distribute them randomly across the screen. Text files added of herbivores at a number of stages of development.

Creatures initialised from a base creature have energy set by start energy percentage. Creatures initialised with variables have their start energy set as an attribute (taken as a percntage of the parents). Max energy increased by 10x, start energy percentage unchanged.

When a carnivore tries to move into a space that is occupied by a herbivore it takes all the herbivores energy and kills it. In the replicate function creatures offspring now have a 10% chance of switching between herbivore and carnivore.

Added creature decision "if cell direction is free".
Added creature decision "compare carnivore direction rank".

Creatures no longer immortal, they have a small chance to die every tick that increases linearly with age.

Notes:
When spawning developed herbivores from a file set the map fill percentage to 35%
While carnivores can appear and sometimes seems to create small groups, it seems very unlikely they will catch on to be a stable species with the current setup.