Added seed to GUI, crude restart, minor tweaks and tidying to game.cpp
The seed now displays in the GUI and can be altered with proper filtering. Placement is still crude.

Added a crude restart, deletes creatureList resourceMap and creatureMap and calls gameInit() again. Often cuases crashes when deleting creatureList. gameInit() needs reworked and restart to avoid this and unneccessary freeing and reallocation of memory that is currently leading to fragmentation.