Combined resource and creature map, changed plant spawn Joined resource and creature map together with each element now a struct of creature pointer and plant value. Has made a significant improvement to performace.
Changed plant spawning. Previously each empty cell looked around for how many mature cells around it then spawned if a threshold was met. Changed to have mature plants spawn seedlings, removes having to look at the 25 adjacent squares but at the cost of not being able to use a buffer array for plants any more. Seedlings that are added ahead of the update path spawn and grow in the same update. not a huge problem but now more strongly a linear program, less ability to multithread.